Add VoltageLevel::TopologyTraverser and Terminal::traverse (#230) (#305)
